WIP 2021-10-28
More Clang compilation fixes.

+++ System/ABI/Integer/Source/System/ABI/Integer/negti.cc
@@ -3,6 +3,6 @@
 #if __SYSTEM_ABI_INTEGER_PROVIDE_128BIT
 __int128_t __negti2(__int128_t i)
 {
-    return System::ABI::Integer::negate(__uint128_t(i));
+    return System::ABI::Integer::negate(i);
 }
 #endif
